+++ st2actions/st2actions/cmd/actionrunner.py
@@ -7,7 +7,6 @@ from st2actions import scheduler, worker
 from st2common import log as logging
 from st2common.service_setup import setup as common_setup
 from st2common.service_setup import teardown as common_teardown
-from st2common.triggers import register_internal_trigger_types
 
 LOG = logging.getLogger(__name__)
 
@@ -23,8 +22,6 @@ eventlet.monkey_patch(
 def _setup():
     common_setup(service='actionrunner', config=config, setup_db=True, register_mq_exchanges=True,
                  register_signal_handlers=True)
-
-    register_internal_trigger_types()
 
 
 def _run_worker():