+++ CHANGELOG.rst
@@ -28,6 +28,8 @@ in development
 * Single sensor mode of Sensor Container uses ``--sensor-ref`` instead of ``--sensor-name``.
 * ``six`` library is now available by default in the Python sandbox to all the newly installed
   packs. (improvement)
+* Dispatch an internal trigger when a datastore item has been created, updated, deleted and when
+  it's value has changed. (new-feature)
 
 0.11.3 - June 16, 2015
 ----------------------